Use snapshot build of kotest
We're switching to the snapshot build of kotest because we need generators for `UInt`, which is not provided in 4.6.0, but is in the snapshot builds.

We could alternatively just use an `Int` generator and convert the result to a `UInt`, but that would  produce a lot of boilerplate, so I've opted to just use the snapshot build for now.