Fix faulty regex pattern for SemVer identifiers
The `IDENTIFIER_CHARACTERS` regex was only expecting 1 character set, which means that any identifier with more than one character would fail throw an exception when the `of` function is used.

The pattern now expects 1 or more characters, as an identifier should not be empty. Empty identifiers gets a special exception thrown, so they're handled by themselves already.